% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering_Deseq2.R
\name{data_filtering_Deseq2}
\alias{data_filtering_Deseq2}
\title{Function to filter DeSeq object}
\usage{
data_filtering_Deseq2(DESeqData)
}
\arguments{
\item{DESeqData}{Deseq object obtained after using Deseg function}
}
\value{
filtered Deseq object
}
\description{
Function to filter DeSeq object
}
\examples{
\dontrun{
dds <- data_filtering_Deseq2(dds)
 }
}
