% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering_EdgeR.R
\name{data_filtering_EdgeR}
\alias{data_filtering_EdgeR}
\title{Function to filter EdgeR count matrix}
\usage{
data_filtering_EdgeR(count_matrix)
}
\arguments{
\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}
}
\value{
Filtered count matrix
}
\description{
Function to filter EdgeR count matrix
}
\examples{
\dontrun{
data1 <- data_filtering_EdgeR(count_matrix)
}
}
