% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization_edgeR.R
\name{data_normalization_edgeR}
\alias{data_normalization_edgeR}
\title{Function to normalize filtered count matrix for EdgeR analysis}
\usage{
data_normalization_edgeR(filtered_count_matrix)
}
\arguments{
\item{filtered_count_matrix}{R dataframe containing filtered gene counts with
genenames as rownames and the sample subset selected for the analysis
as column names}
}
\value{
Normalized count matrix
}
\description{
Function to normalize filtered count matrix for EdgeR analysis
}
\examples{
\dontrun{
d_ss <- data_normalization_edgeR(data1)
}
}
