% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pipeline_input_from_input_file.R
\name{extract_pipeline_input_from_input_file}
\alias{extract_pipeline_input_from_input_file}
\title{Function to extracts input from the pipeline}
\usage{
extract_pipeline_input_from_input_file(arguments_file)
}
\arguments{
\item{arguments_file}{file containing all arguments}
}
\value{
set of inputs for the pipeline
}
\description{
Function to extracts input from the pipeline
}
\examples{
\dontrun{
pipeline_input<-extract_pipeline_input_from_input_file("arguments_file")
}
}
