% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEG_genes.R
\name{find_DEG_genes}
\alias{find_DEG_genes}
\title{Finding the significant from the results}
\usage{
find_DEG_genes(result, p_val, FDR_, LFC_)
}
\arguments{
\item{result}{result obtained from fitting EdgeR or DESeq2}

\item{p_val}{float variable for P_value to be filtered}

\item{FDR_}{float variable for False discovery rate (p adjusted value) to be choosen}

\item{LFC_}{float variable for threshold LogFC value for  regulated genes}
}
\value{
list of signicant genes
}
\description{
Finding the significant from the results
}
