% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_data_DREAM.R
\name{fitting_data_DREAM}
\alias{fitting_data_DREAM}
\title{Function to fit the analysis using DREAM method}
\usage{
fitting_data_DREAM(
  count_matrix,
  meta_data,
  model_design_parameter,
  form,
  contrast_var,
  patient_id_var
)
}
\arguments{
\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}

\item{meta_data}{R dataframe containing information about each sample
with subset samples selected for the analysis as rownames and the information
about samples like the day, participant ID as column names}

\item{model_design_parameter}{string variable used to define
the model design parameter to be used to run the analysis.
for example "~day+ptid", note: it should be the same variable or combination
of variable used to define the contrast variable and patient id variable
for the analysis}

\item{form}{string variable used to define the variable to be tested
for a fixed effect e.g. "~ day + (1/ptid) "}

\item{contrast_var}{string variable used to define the contrast variable
which is the time point variable for the analysis, for example "day"}

\item{patient_id_var}{string variable used to define the unique patient id
variable used for the analysis, for example in the vaccine study it is the
variable that stores participant id for each individual like "ptid"}
}
\value{
DREAM fitting data result
}
\description{
Function to fit the analysis using DREAM method
}
\examples{
 \dontrun{
result_drem <- fitting_data_DREAM (count_matrix = counts,
meta_data = sample_meta_data,
model_design_parameter = "~day",
form = "~ day + (1|pubid) ",
contrast_var="day",
patient_id_var="pubid")
}

}
