% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_data_kimma.R
\name{fitting_data_kimma}
\alias{fitting_data_kimma}
\title{fitting data using kimma method}
\usage{
fitting_data_kimma(
  count_matrix,
  meta_data,
  model_design_parameter,
  kimma_model,
  contrast_var,
  patient_id_var
)
}
\arguments{
\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}

\item{meta_data}{R dataframe containing information about each sample
with subset samples selected for the analysis as rownames and the information
about samples like the day, participant ID as column names}

\item{model_design_parameter}{string variable used to define
the model design parameter to be used to run the analysis.
for example "~day+ptid", note: it should be the same variable or combination
of variable used to define the contrast variable and patient id variable
for the analysis}

\item{kimma_model}{string variable  used to define the model design
parameter used to perform the kimma analysis e.g "day+ptid"}

\item{contrast_var}{string variable used to define the contrast variable
which is the time point variable for the analysis, for example "day"}

\item{patient_id_var}{string variable used to define the unique patient id
variable used for the analysis, for example in the vaccine study it is the
variable that stores participant id for each individual like "ptid"}
}
\value{
result of kimma fitted data
}
\description{
fitting data using kimma method
}
\examples{
\dontrun{
result <- fitting_data_kimma(count_matrix = counts,
meta_data = sample_meta_data,
model_design_parameter = "~day",
kimma_model=  "~ day + (1|pubid)",
contrast_var="day",
patient_id_var="pubid")
}

}
