% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_top_100_deg_Deseq2.R
\name{heatmap_top_100_deg_Deseq2}
\alias{heatmap_top_100_deg_Deseq2}
\title{heatmap plot for DESeq2}
\usage{
heatmap_top_100_deg_Deseq2(
  result,
  count_matrix,
  meta_data,
  model_design_parameter
)
}
\arguments{
\item{result}{result obtained from fitting the Deseq data}

\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}

\item{meta_data}{R dataframe containing information about each sample
with subset samples selected for the analysis as rownames and the information
about samples like the day, participant ID as column names}

\item{model_design_parameter}{string variable used to define
the model design parameter to be used to run the analysis.
for example "~day+ptid", note: it should be the same variable or combination
of variable used to define the contrast variable and patient id variable
for the analysis}
}
\value{
heatmap plot for DESeq2
}
\description{
heatmap plot for DESeq2
}
