% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_top_100_deg_edgeR.R
\name{heatmap_top_100_deg_edgeR}
\alias{heatmap_top_100_deg_edgeR}
\title{heatmap plot for EdgeR}
\usage{
heatmap_top_100_deg_edgeR(result, count_matrix)
}
\arguments{
\item{result}{result obtained from fitting the EdgeR data}

\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}
}
\value{
heatmap plot for EdgeR
}
\description{
heatmap plot for EdgeR
}
