% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_plot_p_value.R
\name{histogram_plot_p_value}
\alias{histogram_plot_p_value}
\title{histogram plot for p_value and adjusted p_value
 for result of fitting data}
\usage{
histogram_plot_p_value(res)
}
\arguments{
\item{res}{result obtained from fitting EdgeR or DESeq2}
}
\value{
histogram plot
}
\description{
histogram plot for p_value and adjusted p_value
 for result of fitting data
}
