% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_DEG.R
\name{implement_DEG}
\alias{implement_DEG}
\title{Function to implement all different kind of analysis}
\usage{
implement_DEG(
  method,
  count_matrix,
  meta_data,
  model_design_parameter,
  form,
  kimma_model,
  contrast_var,
  patient_id_var
)
}
\arguments{
\item{method}{string to choose the analysis to run : limma_voom, limma_voom_duplicate_corr,
deseq}

\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}

\item{meta_data}{R dataframe containing information about each sample
with subset samples selected for the analysis as rownames and the information
about samples like the day, participant ID as column names}

\item{model_design_parameter}{string variable used to define
the model design parameter to be used to run the analysis.
for example "~day+ptid", note: it should be the same variable or combination
of variable used to define the contrast variable and patient id variable
for the analysis}

\item{form}{string variable used to define the variable to be tested
for a fixed effect e.g. "~ day + (1/ptid) "}

\item{kimma_model}{string variable  used to define the model design
parameter used to perform the kimma analysis e.g "day+ptid"}

\item{contrast_var}{string variable used to define the contrast variable
which is the time point variable for the analysis, for example "day"}

\item{patient_id_var}{string variable used to define the unique patient id
variable used for the analysis, for example in the vaccine study it is the
variable that stores participant id for each individual like "ptid"}
}
\value{
list of Normalized count matrix and fit data result for EdgeR or
DESeq2
}
\description{
Function to implement all different kind of analysis
}
