% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFC_scatter_plot_comparison_up.R
\name{LFC_scatter_plot_comparison_up}
\alias{LFC_scatter_plot_comparison_up}
\title{Scatter plot of LFC value comparision to study the comparison between results
 of two different analysis type for upregulated genes}
\usage{
LFC_scatter_plot_comparison_up(
  res1,
  res2,
  LFC_xlabel,
  LFC_ylabel,
  LFC_title,
  p_val,
  FDR_,
  LFC_
)
}
\arguments{
\item{res1}{results from first analysis}

\item{res2}{results from second analysis}

\item{LFC_xlabel}{string variable for x axis label for log10(p_value) of res1}

\item{LFC_ylabel}{string variable for x axis label for log10(p_value) of res2}

\item{LFC_title}{string variable for title for log10(p_value) scatter comparison plot}

\item{p_val}{float value P_value to be filtered}

\item{FDR_}{float value False discovery rate (p adjusted value) to be choosen}

\item{LFC_}{float value for threshold LogFC value for  regulated genes}
}
\value{
a plot comparing LFC vales of upregulated genes
two different analysis types
}
\description{
Scatter plot of LFC value comparision to study the comparison between results
 of two different analysis type for upregulated genes
}
