% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_plot_deseq.R
\name{maplot_deseq}
\alias{maplot_deseq}
\title{plot MA for Deseq}
\usage{
maplot_deseq(method, result, thresh = 0.2, labelsig = TRUE, textcx = 1)
}
\arguments{
\item{method}{deseq analysis is to be run}

\item{result}{result obtained from fitting the data}

\item{thresh}{log fold change threshold value}

\item{labelsig}{boolean for significant threshold display}

\item{textcx}{size of the display}
}
\value{
MA plot
}
\description{
plot MA for Deseq
}
\examples{
\dontrun{
maplot_deseq("deseq",res)
}
}
