% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_counts_data_voom.R
\name{normalized_counts_data_voom}
\alias{normalized_counts_data_voom}
\title{Function to normalize filtered count matrix for EdgeR analysis}
\usage{
normalized_counts_data_voom(count_matrix)
}
\arguments{
\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}
}
\value{
Normalized count matrix
}
\description{
Function to normalize filtered count matrix for EdgeR analysis
}
