% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalized_data_plot.R
\name{Normalized_data_plot}
\alias{Normalized_data_plot}
\title{Function to plot normalized data plot}
\usage{
Normalized_data_plot(
  method,
  result,
  count_matrix,
  meta_data,
  model_design_parameter
)
}
\arguments{
\item{method}{string to choose the analysis to run : limma_voom, limma_voom_duplicate_corr,
deseq}

\item{result}{result obtained from fitting data}

\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}

\item{meta_data}{R dataframe containing information about each sample
with subset samples selected for the analysis as rownames and the information
about samples like the day, participant ID as column names}

\item{model_design_parameter}{string variable used to define
the model design parameter to be used to run the analysis.
for example "~day+ptid", note: it should be the same variable or combination
of variable used to define the contrast variable and patient id variable
for the analysis}
}
\value{
normalized data plot
}
\description{
Function to plot normalized data plot
}
