% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalized_data_plot_deseq.R
\name{Normalized_data_plot_deseq}
\alias{Normalized_data_plot_deseq}
\title{Deseq variance normalized visualization}
\usage{
Normalized_data_plot_deseq(count_matrix, meta_data, model_design_parameter)
}
\arguments{
\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}

\item{meta_data}{R dataframe containing information about each sample
with subset samples selected for the analysis as rownames and the information
about samples like the day, participant ID as column names}

\item{model_design_parameter}{string variable used to define
the model design parameter to be used to run the analysis.
for example "~day+ptid", note: it should be the same variable or combination
of variable used to define the contrast variable and patient id variable
for the analysis}
}
\value{
boxplot of variance normalized DESeq data plot
}
\description{
Deseq variance normalized visualization
}
