% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalized_data_plot_edgeR.R
\name{Normalized_data_plot_edgeR}
\alias{Normalized_data_plot_edgeR}
\title{TMM and voom normalized visualization}
\usage{
Normalized_data_plot_edgeR(result, count_matrix)
}
\arguments{
\item{result}{result obtained by fitting the data}

\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}
}
\value{
boxplot of normalized EdgeR data plot
}
\description{
TMM and voom normalized visualization
}
