% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot_Deseq.R
\name{PCA_plot_Deseq}
\alias{PCA_plot_Deseq}
\title{plot PCA for DESeq2}
\usage{
PCA_plot_Deseq(
  method,
  count_matrix,
  meta_data,
  model_design_parameter,
  pca_group
)
}
\arguments{
\item{method}{Deseq analysis}

\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}

\item{meta_data}{R dataframe containing information about each sample
with subset samples selected for the analysis as rownames and the information
about samples like the day, participant ID as column names}

\item{model_design_parameter}{string variable used to define
the model design parameter to be used to run the analysis.
for example "~day+ptid", note: it should be the same variable or combination
of variable used to define the contrast variable and patient id variable
for the analysis}

\item{pca_group}{groups to be used to check the PCA plot}
}
\value{
pca plot
}
\description{
plot PCA for DESeq2
}
\examples{
\dontrun{
pca_group <- c("day", "Treatment_Group")
PCA_plot_Deseq(Deseq,res,pca_group)
}
}
