% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{plot_mds}
\alias{plot_mds}
\title{plot MDS for EdgeR data fitting}
\usage{
plot_mds(method, result, count_matrix, meta_data, contrast_var, mds_title)
}
\arguments{
\item{method}{anaysis method to be chosen}

\item{result}{result obtained from fitting data}

\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}

\item{meta_data}{R dataframe containing information about each sample
with subset samples selected for the analysis as rownames and the information
about samples like the day, participant ID as column names}

\item{contrast_var}{string variable used to define the contrast variable
which is the time point variable for the analysis, for example "day"}

\item{mds_title}{title for the plot}
}
\value{
plot MDS
}
\description{
plot MDS for EdgeR data fitting
}
