% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normalized_data_voom.R
\name{plot_normalized_data_voom}
\alias{plot_normalized_data_voom}
\title{Function to normalize filtered count matrix for EdgeR analysis}
\usage{
plot_normalized_data_voom(count_matrix)
}
\arguments{
\item{count_matrix}{R dataframe containing gene counts with genenames as
rownames and the sample subset selected for the analysis as column names}
}
\value{
Normalized count matrix boxplot
}
\description{
Function to normalize filtered count matrix for EdgeR analysis
}
