% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Print Venn diagram for Up regulated and down downregulated gene}
\usage{
plot_venn(sig_res, sig_res2, sig_res_category)
}
\arguments{
\item{sig_res}{significant gene results for first comaparision}

\item{sig_res2}{significant gene results for second comaparision}

\item{sig_res_category}{string variable for the title for the signicant gene results
for both comaprision}
}
\value{
venn plot
}
\description{
Print Venn diagram for Up regulated and down downregulated gene
}
\examples{
\dontrun{
 sig_res<-sig_result_0_3
 sig_res2<-sig_result_0_3_trtment_grp
 sig_res_category <- c("0 vs 3", "0_3_trtment_grp")
 plot_venn(sig_res,sig_res2,sig_res_category)
}

}
