% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot for EdgeR or DESeq2 analysis}
\usage{
volcano_plot(result, plot_title)
}
\arguments{
\item{result}{result obtained from fitting data}

\item{plot_title}{string containing title of the plot}
}
\value{
volcano plot
}
\description{
Volcano plot for EdgeR or DESeq2 analysis
}
