% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{function to write the fit data output to a folder}
\usage{
write_output(pipeline_input, res, method)
}
\arguments{
\item{pipeline_input}{piprline inputs to be used}

\item{res}{result obtained after fitting the dat set}

\item{method}{the analysis method choosen}
}
\value{
csv file containing the output
}
\description{
function to write the fit data output to a folder
}
